+++ .github/workflows/ci.yml
@@ -48,20 +48,9 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - name: Install backend
-        uses: cypress-io/github-action@v6
-        with:
-          install-command: npm run install:backend
-          runTests: false
-      - name: Install frontend
-        uses: cypress-io/github-action@v6
-        with:
-          install-command: npm run install:frontend
-          runTests: false
       - name: Run Components tests
         uses: cypress-io/github-action@v6
         with:
-          start: npm run start:frontend, npm run start:server
           command: npm run component:run:cloud
         env:
           CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
@@ -78,15 +67,10 @@ jobs:
         with:
           install-command: npm run install:backend
           runTests: false
-      - name: Install frontend
-        uses: cypress-io/github-action@v6
-        with:
-          install-command: npm run install:frontend
-          runTests: false
       - name: Run API tests
         uses: cypress-io/github-action@v6
         with:
-          start: npm run start:frontend, npm run start:server
+          start: npm run start:server
           command: npm run api:run:cloud
         env:
           CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
