+++ cypress.config.js
@@ -1,26 +1,20 @@
 const { defineConfig } = require('cypress')
 
 module.exports = defineConfig ({
-
   projectId: "kvars8",
-  
   viewportWidth: 1920,
   viewportHeight: 1080,
-
   env: {
     API_URL: "http://localhost:3001",
   },
-
   e2e: {
     baseUrl: "http://localhost:3000",
     retries: {
       runMode: 1,
       openMode: 0,
     },
-
     experimentalRunAllSpecs: true,
   },
-
   component: {
     devServer: {
       framework: 'create-react-app',
