+++ cypress/fixtures/allCustomers.json
@@ -165,3 +165,4 @@
       "totalCustomers": 10
     }
   }
+
