+++ cypress/fixtures/allSizeCustomers.json
@@ -86,3 +86,4 @@
       "totalCustomers": 5
     }
   }
+
