+++ cypress/fixtures/csv
@@ -1,0 +1,1 @@
+ID,Company_Name,Number_of_Employees,Size,Contact_Name,Contact_Email,Street,City,State,Zip_Code,Country\n"1","Jacobs Co","99","Small","","","988 Kimberly Fort Apt. 921","Lake Tracy","Connecticut","07115","United States of America"\n"2","Kilback Co","100","Medium","Daija","Daija_Gislason93@gmail.com","5099 Murray Inlet","South Tiffany","Kentucky","08496","United States of America"\n"3","Parisian Co","999","Medium","Alysson","Alysson.Lang@hotmail.com","43247 Bennett Keys Apt. 999","New Paulside","Connecticut","87855","United States of America"\n"4","Wilderman Co","1000","Enterprise","Brando","Brando_Kozey48@gmail.com","8643 Jackson Wall","Lake Davidstad","Minnesota","29481","United States of America"\n"5","Runolfsson Co","9999","Enterprise","","","851 John Shores Suite 956","New Mariah","Ohio","78314","United States of America"\n"6","Littel Co","10000","Large Enterprise","Selena","Selena.Gleichner7@gmail.com","14135 Kari Garden Suite 427","Mooreshire","Nevada","64043","United States of America"\n"7","Weber Co","49999","Large Enterprise","Malika","Malika16@hotmail.com","70738 Mike Rue","Whitechester","Kentucky","57787","United States of America"\n"8","Lowe Co","50000","Very Large Enterprise","","","87908 Adkins Islands Apt. 944","West Sarah","Georgia","79943","United States of America"\n"9","Waelchi, Wuckert and Koss","85836","Very Large Enterprise","Kellen","Kellen_Ankunding64@hotmail.com","284 Dickens Spring","South Antonia","Kentucky","11853-9574","United States of America"\n"10","Mitchell, Harber and Lynch","51581","Very Large Enterprise","Vella","Vella55@hotmail.com","78158 Pearl Street","New Orrinbury","Rhode Island","37629","United States of America
