+++ cypress/fixtures/enterpriseCustomers.json
@@ -24,3 +24,4 @@
       "totalCustomers": 1
     }
   }
+
