+++ cypress/fixtures/largeEnterpriseCustomers.json
@@ -58,3 +58,4 @@
       "totalCustomers": 3
     }
   }
+
