+++ cypress/fixtures/smallCustomers.json
@@ -77,3 +77,4 @@
       "totalCustomers": 5
     }
   }
+
