+++ cypress/fixtures/veryLargeEnterpriseCustomers.json
@@ -142,3 +142,4 @@
       "totalCustomers": 9
     }
   }
+
