+++ cypress/support/commands.js
@@ -23,14 +23,3 @@
 //
 // -- This will overwrite an existing command --
 // Cypress.Commands.overwrite('visit', (originalFn, url, options) => { ... })
-
-Cypress.Commands.add('noCustomerAvailable', () => {
-    cy.intercept(
-      'GET',
-      `${Cypress.env('API_URL')}/customers**`,
-      { body: '' }
-    ).as('getEmptyCustomers');
-  
-    cy.visit('/');
-    cy.wait('@getEmptyCustomers');
-  });
