+++ frontend/cypress/support/component.js
