+++ package.json
@@ -10,14 +10,14 @@
     "cy:open": "cypress open",
     "cy:run": "cypress run",
     "lint": "eslint . --ignore-pattern 'backend/*' --ignore-pattern 'frontend/*'",
-    "api:run:cloud": "cypress run --record --tag 'API' --spec cypress/e2e/api-tests/**.cy.js",
-    "api:run": "cypress run --spec cypress/e2e/api-tests/**.cy.js",
-    "gui:run:cloud": "cypress run --record --tag 'GUI' --spec cypress/e2e/gui-tests/**.cy.js",
+    "api:run:cloud": "cypress run --record --tag 'API' '--spec cypress/e2e/api-tests/**.cy.js'",
+    "api:run": "cypress run --spec 'cypress/e2e/api-tests/**.cy.js'",
+    "gui:run:cloud": "cypress run --record --tag 'GUI' --spec 'cypress/e2e/gui-tests/**.cy.js'",
     "gui:run": "cypress run --spec cypress/e2e/gui-tests/**.cy.js",
     "component:run:cloud": "cypress run --record  --tag 'component' --component",
     "component:run": "cypress run --component",
-    "accessibility:run": "cypress run --record --tag 'a11y' --spec cypress/e2e/accessibility-tests/**.cy.js",
-    "cloud:run": "npx cypress run && npm run --component"  
+    "accessibility:run": "cypress run --record --tag 'a11y' --spec 'cypress/e2e/accessibility-tests/**.cy.js'",
+    "cloud:run": "cypress run && npm run --component"  
   },
 
   "repository": {
